feat: inner class to upper level (#1)
